+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.4.1 - 6/29/2015
+
+- Fix #16: Errror provisioning chromedriver on centos7
+- Fix #15: Drivers not copied to /selenium/drivers/ folders on Windows 7
+
 ## 2.4.0 - 6/9/2015
 
 - Allow resources to be globally configured